Tweak sleep test assertion
Change ZestActionSleepUnitTest to use an assertion with delta when
checking time slept, also, swap the values (to have expected value and
actual value in the expected positions) and extract the sleep time into
a variable.
The reason to do the change was that the test failed:
   AssertionError: expected:<201> but was:<200>
while it was still within the expected margin.