Fix loop's statements copy
Change ZestLoop to return the new list, when copying the statements, not
the same list that's being copied, also, return an empty list instead of
null when empty/null (does not require calling code to handle the null
case). As part of the change the new list was renamed to not clash with
the instance variable.
Add test to assert the expected behaviour.