Change action invoke to allow to set the charset
Change ZestActionInvoke to allow to set the charset used in read
operations, to ensure the script is loaded or the output of the program
read, always with expected charset (if not it would rely on the JVM
default charset, which is not always the same).
Update tests to check that the charset is de/serialised.

Related to zaproxy/zaproxy#3377 - ZAP can no longer load non UTF8
scripts