Accept insecure certs when launching browsers
Change ZestClientLaunch to accept "insecure" certificates when launching
the browsers, to improve the interoperability with sites and other tools
(for example, ZAP which uses a custom Root CA certificate).
The following browsers are accepting the "insecure" certificates:
Chrome, Firefox, HtmlUnit, JBD, and PhantomJS.

Chrome and HtmlUnit were already accepting the insecure certs. For
Chrome it was disabled the headless mode because of an issue that
prevents the certs from being accepted when in headless mode.