Allow to obtain a screenshot from the browsers
Add a ZestClient action, ZestClientScreenshot, that allows to take a
screenshot from a browser (e.g Chrome, Firefox), to allow to obtain a
visual evidence of a possible issue (e.g. DOM XSS).
Add tests to assert the expected behaviour.
Change ZestClientFailException to allow to specify a message and the
cause of the exception.
Add Mockito 2.13.0 as test dependency, to ease the tests.